v2.1.0: modernize, fix BufferedRadnomProvider typo, upgrade test stack, add CI/CD
- Phase 4: add BufferedRandomProvider (fixed spelling); keep BufferedRadnomProvider as
  [Obsolete] shim for backwards compatibility; rename file and test class to match
- Phase 3: remove stale Resources.resx/Designer.cs refs from csproj; add SourceLink
  8.0.0, LangVersion latest, Nullable enable, GenerateDocumentationFile,
  PackageReadmeFile; remove GeneratePackageOnBuild/PackageLicenseUrl/PackageReleaseNotes;
  update copyright 2018→2026; bump v2.0.0→2.1.0
- Test: netcoreapp2.0→net8.0; NSubstitute 3.1.0→5.3.0; xunit 2.3.1→2.9.2;
  TestSdk 15.5.0→17.12.0; add coverlet; fix nullable warnings; fix unused theory param
- Phase 6: add ci.yml, publish.yml (OIDC Trusted Publishing), sonar.yml (with coverage)
- README: full rewrite with CI/Sonar/NuGet/license/BMC badges, API table, usage examples
- .gitattributes: *.yml text eol=lf
